requirements.txt
Flask==2.2.5
numpy==1.21.5
Pillow==9.0.1
Pillow==10.1.0
scipy==1.7.3
torch==1.13.1
torchvision==0.14.1
